requirements.txt
Django
PyQt5
PyQtWebEngine
pyserial
requests